requirements.txt
appdirs==1.4.4
async-timeout==4.0.2
beautifulsoup4==4.11.1
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-redis==0.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
frozendict==2.3.4
greenlet==2.0.1
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multitasking==0.0.11
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
tenacity==8.1.0
urllib3==1.26.14
webencodings==0.5.1
Werkzeug==2.2.2
WTForms==3.0.1
yfinance==0.2.9